clear parsing.__init__
